C++ coding within various game engine classes
Entity contains an abstract base class, with various sub-classes for other entity types created.
Image contains various functions to handle the loading, drawing and clipping of sprites/sprite-sheets, with or without transparency. 
World contains various functions to handle the initalising of the engine, with access to data using a singleton pattern and black boxed classes.